+++ sw-sysemu/memory/dense_region.h
@@ -16,7 +16,7 @@ namespace bemu {
 extern typename MemoryRegion::value_type memory_reset_value;
 
 
-template<unsigned long long Base, size_t N, bool Writeable=true>
+template<unsigned long long Base, unsigned long long N, bool Writeable=true>
 struct DenseRegion : public MemoryRegion {
     typedef typename MemoryRegion::addr_type      addr_type;
     typedef typename MemoryRegion::size_type      size_type;
